fix: sorting behaviour
Figured out what the problem was with sorting, because I was reversing
the whole array when the toggle flipped up, the order the notes where
sorted in also reversed. Which didn't make a difference to the notes
whos' order was effected by sort(), but any that *weren't* supposed to
be effected by sort() reversed order.