dev: prevent entering target date if current input value is invalid
partly implemented. for now it blocks entering input when the entire
value is not empty, and the value is greater than the current date.
unfortunately as soon as the user starts filling in the year the
value is considered not empty, but the year start filling up from the
last digit so no matter what the value will be invalid.

should be simple enough to fix this, if in a hacky kinda way